Generalize update-lots.js into base.js
