+++ scripts/tests/data/empty-update.geojson
