+++ scripts/tests/data/valid-update.geojson
